# backend/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.Categoria:
    properties:
      archivada:
        type: boolean
      color:
        type: string
      icono:
        type: string
      id:
        type: string
      nombre:
        type: string
      tipo:
        type: string
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCategoria:
    properties:
      archivada:
        type: boolean
      color:
        type: string
      icono:
        maxLength: 40
        type: string
      nombre:
        maxLength: 60
        minLength: 1
        type: string
      tipo:
        enum:
        - ingreso
        - gasto
        type: string
    required:
    - color
    - icono
    - nombre
    - tipo
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.ResultadoImportacion:
    properties:
      importadas:
        type: integer
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle:
    properties:
      codigo:
        type: string
      detalles:
        items:
          type: string
        type: array
      mensaje:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta:
    properties:
      limite:
        type: integer
      pagina:
        type: integer
      total:
        type: integer
      total_paginas:
        type: integer
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre:
    properties:
      datos: {}
      meta:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta'
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError:
    properties:
      error:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API REST de finanzas personales: cuentas, categorias, transacciones, presupuestos y reportes.
    Todos los datos pertenecen a un usuario. El identificador del dueño sale siempre del token, nunca del cuerpo de la peticion.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinTrack API
  version: 0.1.0
paths:
...
  /categorias:
    get:
      description: Solo las del usuario del token.
      parameters:
      - description: Filtra por tipo
        enum:
        - ingreso
        - gasto
        in: query
        name: tipo
        type: string
      - description: Incluye tambien las archivadas
        in: query
        name: incluir_archivadas
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Categoria'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Listar categorias
      tags:
      - categorias
    post:
      parameters:
      - description: Datos de la categoria
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCategoria'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Categoria'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Crear una categoria
      tags:
      - categorias
  /categorias/{id}:
    delete:
      description: Responde 409 si tiene movimientos o presupuestos.
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Sin contenido
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Borrar una categoria
      tags:
      - categorias
    get:
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Categoria'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Obtener una categoria
      tags:
      - categorias
    put:
      description: No se puede cambiar el tipo de una categoria que ya tiene movimientos
        (409).
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      - description: Datos de la categoria
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCategoria'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Categoria'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Editar una categoria
      tags:
      - categorias
...
  /transacciones/exportar:
    get:
      description: Descarga un archivo con los mismos filtros del listado, sin paginar.
        Las cuentas y categorias salen por nombre.
      parameters:
      - description: Fecha inicial AAAA-MM-DD
        in: query
        name: desde
        type: string
      - description: Fecha final AAAA-MM-DD, incluida
        in: query
        name: hasta
        type: string
      - description: ingreso o gasto
        enum:
        - ingreso
        - gasto
        in: query
        name: tipo
        type: string
      - description: Filtra por categoria
        in: query
        name: categoria_id
        type: string
      - description: Filtra por cuenta
        in: query
        name: cuenta_id
        type: string
      - description: Texto en descripcion o notas
        in: query
        name: busqueda
        type: string
      - description: Orden del listado
        enum:
        - fecha_desc
        - fecha_asc
        - monto_desc
        - monto_asc
        in: query
        name: orden
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: Archivo CSV
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Exportar transacciones a CSV
      tags:
      - csv
  /transacciones/importar:
    post:
      consumes:
      - multipart/form-data
      description: Sube un CSV con las columnas fecha, tipo, cuenta, categoria, monto,
        descripcion y notas. O entra el archivo completo o no entra nada.
      parameters:
      - description: Archivo CSV, maximo 2 MiB
        in: formData
        name: archivo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.ResultadoImportacion'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Importar transacciones desde CSV
      tags:
      - csv
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Token de acceso con el prefijo. Ejemplo: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"